// src/lib/types.ts
export interface IDesktopStakingCard {
  id: number;
  title: string;
  tokenId: string;
  image: string;
  amount: string;
  duration: string;
  reward: string;
  percentage: string;
}

export interface IMobileStakingCard {
  id: number;
  title: string;
  tokenId: string;
  image: string;
  amount: string;
  reward: string;
  percentage: string;
}